This update experiments with arrows only being able to be in air for a given amount of time, and if they don't land they disintegrate; one second at the moment.  The thinking behind this is that using proximity to prevent flies from being attracted to faraway arrows may cause problems in the long run; nothing keeps the player from collecting flies and putting them far away from critical points in the map.
This way, it'll be more difficult to get rid of flies, preventing the player from both exploiting the proximity logic and from exploiting flies following arrows off the map.  This also keeps flies closer during parts of the map that are meant to be fast moving, because the player will only be able to distract the flies for so long.